Fixed Issue 62 Bundler Version Out of Date
For this commit I resolved the issue with the original Docker image used having
incorrect dependencies by switching to the official Jekyll Docker image.

This required quickly editing the image specified in the docker-compose.yml
file and adding a command to run on the starting the service to behave as
expected in the README.md.

Will